Fix duplicated functions in newly broken up files
